backend/requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
Authlib==1.5.2
bcrypt==4.0.1
black==25.1.0
certifi==2025.4.26
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isort==6.0.1
itsdangerous==2.2.0
libnacl==2.1.0
Mako==1.3.10
MarkupSafe==3.0.2
mypy==1.15.0
mypy_extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.7
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
PyNaCl==1.5.0
pyotp==2.9.0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
qrcode==8.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
slowapi==0.1.9
pytest==8.3.4
pytest-cov==6.0.0